Refactor:Extract sections
