+++ src/main/java/nextstep/subway/applicaion/dto/LineResponse.java
@@ -38,7 +38,7 @@ public class LineResponse {
     }
 
     private static List<StationResponse> createStationResponses(Line line) {
-        if (line.getSections().isEmpty()) {
+        if (line.getStations().isEmpty()) {
             return Collections.emptyList();
         }
