+++ src/main/java/nextstep/subway/domain/Line.java
@@ -1,13 +1,10 @@
 package nextstep.subway.domain;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Entity
 public class Line extends BaseEntity {
-    private static final int MIN_SIZE = 1;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,8 +13,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -30,7 +27,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Section section) {
         this.name = name;
         this.color = color;
-        this.sections.add(section);
+        this.sections = new Sections(section);
         section.setLine(this);
     }
 
@@ -40,86 +37,12 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(Section section) {
-        if (sections.isEmpty()) {
-            sections.add(section);
-            section.setLine(this);
-            return;
-        }
-
-        Station upStation = section.getUpStation();
-        Station downStation = section.getDownStation();
-        int newDistance = section.getDistance();
-
-        List<Station> stations = getStations();
-
-        boolean hasUpStation = stations.contains(upStation);
-        boolean hasDownStation = stations.contains(downStation);
-
-        if (hasUpStation && hasDownStation) {
-            throw new IllegalArgumentException("이미 등록된 구간입니다.");
-        }
-
-        if (!hasUpStation && !hasDownStation) {
-            throw new IllegalArgumentException("요청한 상/하행역이 모두 노선에 등록되지 않았습니다.");
-        }
-
-        if (hasUpStation) {
-            int index = indexOfUpStation(upStation);
-
-            if(index != -1){
-                Section oldSection = sections.get(index);
-                if(oldSection.getDistance() <= newDistance){
-                    throw new IllegalArgumentException("길이가 더 긴 구간은 추가할 수 없습니다.");
-                }
-                oldSection.updateSection(downStation, oldSection.getDownStation(), oldSection.getDistance() - newDistance);
-            }
-        }
-
-        if(hasDownStation) {
-            int index = indexOfDownStation(downStation);
-
-            if(index != -1){
-                Section oldSection = sections.get(index);
-                if(oldSection.getDistance() <= newDistance){
-                    throw new IllegalArgumentException("길이가 더 긴 구간은 추가할 수 없습니다.");
-                }
-                oldSection.updateSection(oldSection.getUpStation(), upStation, oldSection.getDistance() - newDistance);
-            }
-        }
-
-        sections.add(section);
+        sections.addSection(section);
         section.setLine(this);
     }
 
-    private int indexOfUpStation(Station station) {
-        for (int i = 0; i < sections.size(); i++) {
-            Station upStation = sections.get(i).getUpStation();
-            if (upStation.equals(station)) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
-    private int indexOfDownStation(Station station) {
-        for (int i = 0; i < sections.size(); i++) {
-            Station downStation = sections.get(i).getDownStation();
-            if (downStation.equals(station)) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
-    public void removeStation(Station newStation) {
-        if (sections.size() == MIN_SIZE) {
-            throw new IllegalArgumentException("구간이 1개일 경우 삭제가 불가능합니다.");
-        }
-        List<Station> stations = getStations();
-        if (!stations.get(stations.size() - 1).equals(newStation)) {
-            throw new IllegalArgumentException("마지막 역만 삭제가 가능합니다.");
-        }
-        sections.remove(sections.size() - 1);
+    public void removeStation(Station station){
+        sections.removeStation(station);
     }
 
     public Long getId() {
@@ -135,72 +58,10 @@ public class Line extends BaseEntity {
     }
 
     public List<Section> getSections() {
-        return sections;
+        return sections.getSections();
     }
 
     public List<Station> getStations() {
-        List<Section> sorted = sort();
-        List<Station> stations = sorted.stream()
-                .map(Section::getUpStation)
-                .collect(Collectors.toList());
-
-        stations.add(sorted.get(sorted.size() - 1).getDownStation());
-
-        return stations;
-    }
-
-    private List<Section> sort() {
-        List<Section> sorted = new ArrayList<>();
-        Station first = findFirst();
-        Station last = findLast();
-
-        Station now = first;
-
-        do {
-            Section section = find(now);
-            sorted.add(section);
-            now = section.getDownStation();
-        } while (!now.equals(last));
-
-        return sorted;
-    }
-
-    private Section find(Station station) {
-        return sections.stream()
-                .filter(section -> station.equals(section.getUpStation()))
-                .findFirst()
-                .get();
-    }
-
-    private Station findFirst() {
-        List<Station> upStations = getUpStations();
-        List<Station> downStations = getDownStations();
-
-        return upStations.stream()
-                .filter(station -> !downStations.contains(station))
-                .findFirst()
-                .get();
-    }
-
-    private Station findLast() {
-        List<Station> upStations = getUpStations();
-        List<Station> downStations = getDownStations();
-
-        return downStations.stream()
-                .filter(station -> !upStations.contains(station))
-                .findFirst()
-                .get();
-    }
-
-    private List<Station> getUpStations() {
-        return sections.stream()
-                .map(Section::getUpStation)
-                .collect(Collectors.toList());
-    }
-
-    private List<Station> getDownStations() {
-        return sections.stream()
-                .map(Section::getDownStation)
-                .collect(Collectors.toList());
+        return sections.getStations();
     }
 }
